% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymirt.R
\name{glance.SingleGroupClass}
\alias{glance.SingleGroupClass}
\title{Summarize useful model information, for fitted mirt model}
\usage{
\method{glance}{SingleGroupClass}(x, ...)
}
\arguments{
\item{x}{mirt model (often returned by mirt() function)}

\item{...}{ignored}
}
\value{
tibble with columns for TODO
}
\description{
This function uses mirt's extract.mirt to pull out measures, so
see that function for other measures you could get.
}
\examples{
# We'll use examples from the mirt library (see ?mirt)
library(mirt)

#load mirt library LSAT section 7 data
data <- expand.table(LSAT7)

# one-factor model
mod1 <- mirt(data, 1)
glance(mod1)

# two-factor model
mod2 <- mirt(data, 2)
glance(mod2)

}
