% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymirt.R
\name{tidy.SingleGroupClass}
\alias{tidy.SingleGroupClass}
\title{Convert MIRT SingleGroupClass to tidy tibble}
\usage{
\method{tidy}{SingleGroupClass}(x, group_pars = TRUE, ...)
}
\arguments{
\item{x}{mirt model (often returned by mirt() function)}

\item{group_pars}{whether to include non-item parameters (e.g. latent mean or covariance)}

\item{...}{passed along with model to coef method}
}
\value{
tibble with columns for item_number, item, term (e.g. a1, d), estimate
}
\description{
Convert MIRT SingleGroupClass to tidy tibble
}
\examples{
# We'll use examples from the mirt library (see ?mirt)
library(mirt)

#load mirt library LSAT section 7 data
data <- expand.table(LSAT7)

# one-factor model
mod1 <- mirt(data, 1)
tidy(mod1)
tidy(mod1, group_pars = FALSE)      # only item parameters

# two-factor model
mod2 <- mirt(data, 2)
tidy(mod2)


}
