% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymirt-package.R
\docType{package}
\name{tidymirt-package}
\alias{tidymirt}
\alias{tidymirt-package}
\title{tidymirt: Tidyverse friendly functions for Item Response Theory}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mine-cetinkaya-rundel/tidymirt}
  \item Report bugs at \url{https://github.com/mine-cetinkaya-rundel/tidymirt/issues}
}

}
\author{
\strong{Maintainer}: Mine Çetinkaya-Rundel \email{cetinkaya.mine@gmail.com} (\href{https://orcid.org/0000-0001-6452-2420}{ORCID})

}
\keyword{internal}
