% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_test_versions.R
\name{combine_test_versions}
\alias{combine_test_versions}
\title{Combine two versions of a test}
\usage{
combine_test_versions(test_a, test_b, meta_data)
}
\arguments{
\item{test_a}{A data frame with data from Test A.}

\item{test_b}{A data frame with data from Test A.}

\item{meta_data}{A data frame linking Test A to Test B.}
}
\value{
A combined data frame \code{response_data} and \code{meta_data}.
}
\description{
Combine data from two versions of a test, keeping track of which questions
were unchanged, moved, removed, or added.
}
\examples{

library(tibble)

test_a <- tribble(
  ~A1, ~A2, ~A3, ~A4,
  0,   1,   1,   0,
  0,   0,   1,   1,
  0,   1,   0,   1
)

test_b <- tribble(
  ~Q1, ~Q2, ~Q3, ~Q4,
  1,   1,   1,   0,
  0,   1,   0,   1,
  0,   1,   0,   0
)

meta_data <- tribble(
  ~A,   ~B,
  "A1", "Q1",
  "A2", "Q3",
  "A3", NA,
  "A4", "Q2",
  NA,   "Q4"
)

combine_test_versions(test_a, test_b, meta_data)

}
